- add basic cmake  - move headers to include - add .envrc for local variables
