+++ .envrc
@@ -1,0 +1,2 @@
+export PROJ_NAME="VideoFilterT"
+export VIDEO_NAME="intersection"
