+++ src/filter.cpp
@@ -1,4 +1,4 @@
-#include "filter.h"
+#include "../include/filter.h"
 
 #include <stdint.h>
